32-bit block state ids for FastAsyncWorldEdit 2.15.4
Widens block state ordinals from char to int, adapts the schematic palette and the disk clipboard format, and fixes four registry and adapter bugs seen on NeoForge hybrid cores. Only the 1.21 adapter is built.

+++ worldedit-core/src/main/java/com/fastasyncworldedit/core/extent/filter/LinkedFilter.java
@@ -78,7 +78,7 @@ public sealed class LinkedFilter<L extends Filter, R extends Filter> implements 
         }
 
         @Override
-        public void applyVector(final VectorFacade get, final VectorFacade set, final VectorMask<Short> mask) {
+        public void applyVector(final VectorFacade get, final VectorFacade set, final VectorMask<Integer> mask) {
             getLeft().applyVector(get, set, mask);
             getRight().applyVector(get, set, mask);
         }
